refactor: move config llm residuals internal
